requirements.txt
backports.zoneinfo==0.2.1
dbus-python==1.3.2
pkg-resources==0.0.0
pycairo==1.23.0
PyGObject==3.44.1
